+++ scripts/dev_setup.py
@@ -77,7 +77,7 @@ args = parser.parse_args()
 
 packages = {
     tuple(os.path.dirname(f).rsplit(os.sep, 1))
-    for f in glob.glob("sdk/*/azure-*/setup.py") + glob.glob("tools/azure-*/setup.py")
+    for f in glob.glob(os.path.join(root_dir, "sdk/*/azure-*/setup.py")) + glob.glob(os.path.join(root_dir, "tools/azure-*/setup.py"))
 }
 # [(base_folder, package_name), ...] to {package_name: base_folder, ...}
 packages = {package_name: base_folder for (base_folder, package_name) in packages}