+++ sdk/confluent/azure-mgmt-confluent/azure/mgmt/confluent/models/_models.py
@@ -10,9 +10,17 @@ from azure.core.exceptions import HttpResponseError
 import msrest.serialization
 
 
-class ConfluentAgreementProperties(msrest.serialization.Model):
-    """Terms properties for Marketplace and Confluent.
+class ConfluentAgreementResource(msrest.serialization.Model):
+    """Confluent Agreements Resource.
 
+    Variables are only populated by the server, and will be ignored when sending a request.
+
+    :ivar id: ARM id of the resource.
+    :vartype id: str
+    :ivar name: Name of the agreement.
+    :vartype name: str
+    :ivar type: The type of the resource.
+    :vartype type: str
     :param publisher: Publisher identifier string.
     :type publisher: str
     :param product: Product identifier string.
@@ -32,47 +40,6 @@ class ConfluentAgreementProperties(msrest.serialization.Model):
     :type accepted: bool
     """
 
-    _attribute_map = {
-        'publisher': {'key': 'publisher', 'type': 'str'},
-        'product': {'key': 'product', 'type': 'str'},
-        'plan': {'key': 'plan', 'type': 'str'},
-        'license_text_link': {'key': 'licenseTextLink', 'type': 'str'},
-        'privacy_policy_link': {'key': 'privacyPolicyLink', 'type': 'str'},
-        'retrieve_datetime': {'key': 'retrieveDatetime', 'type': 'iso-8601'},
-        'signature': {'key': 'signature', 'type': 'str'},
-        'accepted': {'key': 'accepted', 'type': 'bool'},
-    }
-
-    def __init__(
-        self,
-        **kwargs
-    ):
-        super(ConfluentAgreementProperties, self).__init__(**kwargs)
-        self.publisher = kwargs.get('publisher', None)
-        self.product = kwargs.get('product', None)
-        self.plan = kwargs.get('plan', None)
-        self.license_text_link = kwargs.get('license_text_link', None)
-        self.privacy_policy_link = kwargs.get('privacy_policy_link', None)
-        self.retrieve_datetime = kwargs.get('retrieve_datetime', None)
-        self.signature = kwargs.get('signature', None)
-        self.accepted = kwargs.get('accepted', None)
-
-
-class ConfluentAgreementResource(msrest.serialization.Model):
-    """Agreement Terms definition.
-
-    Variables are only populated by the server, and will be ignored when sending a request.
-
-    :ivar id: The ARM id of the resource.
-    :vartype id: str
-    :ivar name: The name of the agreement.
-    :vartype name: str
-    :ivar type: The type of the agreement.
-    :vartype type: str
-    :param properties: Represents the properties of the resource.
-    :type properties: ~azure.mgmt.confluent.models.ConfluentAgreementProperties
-    """
-
     _validation = {
         'id': {'readonly': True},
         'name': {'readonly': True},
@@ -83,7 +50,14 @@ class ConfluentAgreementResource(msrest.serialization.Model):
         'id': {'key': 'id', 'type': 'str'},
         'name': {'key': 'name', 'type': 'str'},
         'type': {'key': 'type', 'type': 'str'},
-        'properties': {'key': 'properties', 'type': 'ConfluentAgreementProperties'},
+        'publisher': {'key': 'properties.publisher', 'type': 'str'},
+        'product': {'key': 'properties.product', 'type': 'str'},
+        'plan': {'key': 'properties.plan', 'type': 'str'},
+        'license_text_link': {'key': 'properties.licenseTextLink', 'type': 'str'},
+        'privacy_policy_link': {'key': 'properties.privacyPolicyLink', 'type': 'str'},
+        'retrieve_datetime': {'key': 'properties.retrieveDatetime', 'type': 'iso-8601'},
+        'signature': {'key': 'properties.signature', 'type': 'str'},
+        'accepted': {'key': 'properties.accepted', 'type': 'bool'},
     }
 
     def __init__(
@@ -94,11 +68,18 @@ class ConfluentAgreementResource(msrest.serialization.Model):
         self.id = None
         self.name = None
         self.type = None
-        self.properties = kwargs.get('properties', None)
+        self.publisher = kwargs.get('publisher', None)
+        self.product = kwargs.get('product', None)
+        self.plan = kwargs.get('plan', None)
+        self.license_text_link = kwargs.get('license_text_link', None)
+        self.privacy_policy_link = kwargs.get('privacy_policy_link', None)
+        self.retrieve_datetime = kwargs.get('retrieve_datetime', None)
+        self.signature = kwargs.get('signature', None)
+        self.accepted = kwargs.get('accepted', None)
 
 
 class ConfluentAgreementResourceListResponse(msrest.serialization.Model):
-    """Response of a agreements operation.
+    """Response of a list operation.
 
     :param value: Results of a list operation.
     :type value: list[~azure.mgmt.confluent.models.ConfluentAgreementResource]
@@ -270,11 +251,14 @@ class OperationResult(msrest.serialization.Model):
     :type name: str
     :param display: The object that represents the operation.
     :type display: ~azure.mgmt.confluent.models.OperationDisplay
+    :param is_data_action: Indicates whether the operation is a data action.
+    :type is_data_action: bool
     """
 
     _attribute_map = {
         'name': {'key': 'name', 'type': 'str'},
         'display': {'key': 'display', 'type': 'OperationDisplay'},
+        'is_data_action': {'key': 'isDataAction', 'type': 'bool'},
     }
 
     def __init__(
@@ -284,6 +268,7 @@ class OperationResult(msrest.serialization.Model):
         super(OperationResult, self).__init__(**kwargs)
         self.name = kwargs.get('name', None)
         self.display = kwargs.get('display', None)
+        self.is_data_action = kwargs.get('is_data_action', None)
 
 
 class OrganizationResource(msrest.serialization.Model):
@@ -530,7 +515,7 @@ class UserDetail(msrest.serialization.Model):
     _validation = {
         'first_name': {'max_length': 50, 'min_length': 0},
         'last_name': {'max_length': 50, 'min_length': 0},
-        'email_address': {'pattern': r'^\S+@\S+\.\S+$'},
+        'email_address': {'pattern': r'\S+@\S+\.\S+'},
     }
 
     _attribute_map = {
@@ -563,7 +548,7 @@ class OrganizationResourcePropertiesUserDetail(UserDetail):
     _validation = {
         'first_name': {'max_length': 50, 'min_length': 0},
         'last_name': {'max_length': 50, 'min_length': 0},
-        'email_address': {'pattern': r'^\S+@\S+\.\S+$'},
+        'email_address': {'pattern': r'\S+@\S+\.\S+'},
     }
 
     _attribute_map = {