+++ sdk/eventgrid/azure-eventgrid/samples/champion_scenarios/cs1b_publish_custom_events_to_a_topic_with_signature.py
@@ -20,12 +20,12 @@ from azure.core.credentials import AzureKeyCredential, AzureSasCredential
 from datetime import datetime, timedelta
 
 topic_key = os.environ["EG_ACCESS_KEY"]
-topic_hostname = os.environ["EG_TOPIC_HOSTNAME"]
+endpoint = os.environ["EG_TOPIC_HOSTNAME"]
 expiration_date_utc = datetime.utcnow() + timedelta(hours=1)
 
-signature = generate_shared_access_signature(topic_hostname, topic_key, expiration_date_utc)
+signature = generate_shared_access_signature(endpoint, topic_key, expiration_date_utc)
 credential = AzureSasCredential(signature)
-client = EventGridPublisherClient(topic_hostname, credential)
+client = EventGridPublisherClient(endpoint, credential)
 
 client.send_events([
 	EventGridEvent(