+++ sdk/eventgrid/azure-eventgrid/samples/champion_scenarios/cs2_publish_custom_events_to_a_domain_topic.py
@@ -18,8 +18,8 @@ import os
 from azure.eventgrid import EventGridPublisherClient, EventGridEvent
 from azure.core.credentials import AzureKeyCredential
 
-domain_key = os.environ["EG_ACCESS_KEY"]
-domain_hostname = os.environ["EG_TOPIC_HOSTNAME"]
+domain_key = os.environ["EG_DOMAIN_ACCESS_KEY"]
+domain_hostname = os.environ["EG_DOMAIN_TOPIC_HOSTNAME"]
 
 credential = AzureKeyCredential(domain_key)
 client = EventGridPublisherClient(domain_hostname, credential)