+++ sdk/eventgrid/azure-eventgrid/samples/publish_samples/publish_custom_schema_events_to_topic_sample.py
@@ -25,12 +25,12 @@ from azure.core.credentials import AzureKeyCredential
 from azure.eventgrid import EventGridPublisherClient, CustomEvent
 
 key = os.environ["CUSTOM_SCHEMA_ACCESS_KEY"]
-topic_hostname = os.environ["CUSTOM_SCHEMA_TOPIC_HOSTNAME"]
+endpoint = os.environ["CUSTOM_SCHEMA_TOPIC_HOSTNAME"]
 
 def publish_event():
     # authenticate client
     credential = AzureKeyCredential(key)
-    client = EventGridPublisherClient(topic_hostname, credential)
+    client = EventGridPublisherClient(endpoint, credential)
 
     custom_schema_event = {
         "customSubject": "sample",