+++ sdk/storage/azure-storage-queue/azure/storage/queue/_deserialize.py
@@ -12,7 +12,7 @@ from ._models import QueueProperties
 
 
 def deserialize_metadata(response, obj, headers):
-    raw_metadata = {k: v for k, v in response.headers.items() if k.startswith("x-ms-meta-")}
+    raw_metadata = {k: v for k, v in response.http_response.headers.items() if k.startswith("x-ms-meta-")}
     return {k[10:]: v for k, v in raw_metadata.items()}
 
 
@@ -26,6 +26,7 @@ def deserialize_queue_properties(response, obj, headers):
 
 
 def deserialize_queue_creation(response, obj, headers):
+    response = response.http_response
     if response.status_code == 204:
         error_code = StorageErrorCode.queue_already_exists
         error = ResourceExistsError(