+++ sdk/eventgrid/azure-eventgrid/CHANGELOG.md
@@ -9,6 +9,7 @@
   - `decode_eventgrid_event` is renamed to `deserialize_eventgrid_events`.
   - The system events now exist in the `azure.eventgrid.systemevents` namespace instead of `azure.eventgrid.models` namespace.
   - The `send` method in the `EventGridPubliserClient` is now replaced by the `send_events`.
+  - `topic_hostname` is renamed to `endpoint` in the `EventGridPublisherClient`.
   
   **Bug Fixes**
   - `EventGridEvent` has two additional required positional parameters namely, `data` and `data_version`.